web-app/requirements.txt
python-decouple==3.6
pymongo==4.0.1
Flask==2.1.0
python-dotenv==0.19.2
yattag==1.14.0
google-cloud-datastore==2.4.0
google-auth==2.5.0
requests==2.27.1
dnspython==2.2.1
Flask-PyMongo==2.3.0
google-cloud-storage==2.1.0
gunicorn==20.1.0
werkzeug==2.1.0